refactor: split incident info/update views into standalone modules
