+++ app/modules/incident/incident_helper.py
@@ -3,7 +3,6 @@ import json
 import logging
 import os
 import re
-from boto3.dynamodb.types import TypeDeserializer
 from slack_sdk import WebClient
 from slack_bolt import Ack, Respond, App
 from integrations.google_workspace import google_drive
@@ -17,8 +16,9 @@ from modules.incident import (
     incident_conversation,
     schedule_retro,
     db_operations,
+    information_display,
+    information_update,
 )
-from models.incidents import Incident
 
 INCIDENT_CHANNELS_PATTERN = r"^incident-\d{4}-"
 SRE_DRIVE_ID = os.environ.get("SRE_DRIVE_ID")
@@ -35,6 +35,9 @@ VALID_STATUS = [
 
 
 help_text = """
+\n `/sre incident`
+\n      - opens a modal to show and update the incident information
+\n      - ouvre une boîte de dialogue pour afficher et mettre à jour les informations sur l'incident
 \n `/sre incident create-folder <folder_name>`
 \n      - create a folder for a team in the incident drive
 \n      - créer un dossier pour une équipe dans le dossier d'incidents
@@ -80,7 +83,6 @@ def register(bot: App):
     bot.view("view_save_incident_roles")(incident_roles.save_incident_roles)
     bot.view("view_save_event")(schedule_retro.save_incident_retro)
     bot.action("confirm_click")(schedule_retro.confirm_click)
-    bot.action("update_incident_field")(open_update_field_view)
     bot.action("archive_channel")(incident_conversation.archive_channel_action)
     bot.event("reaction_added", matchers=[incident_conversation.is_floppy_disk])(
         incident_conversation.handle_reaction_added
@@ -95,7 +97,8 @@ def register(bot: App):
         incident_conversation.just_ack_the_rest_of_reaction_events
     )
     bot.view("incident_updates_view")(handle_updates_submission)
-    bot.view("update_field_modal")(handle_update_field_submission)
+    bot.action("update_incident_field")(information_update.open_update_field_view)
+    bot.view("update_field_modal")(information_update.handle_update_field_submission)
 
 
 def handle_incident_command(
@@ -105,7 +108,7 @@ def handle_incident_command(
 
     # If no arguments are provided, open the update status view
     if len(args) == 0:
-        open_incident_info_view(client, body, ack, respond)
+        information_display.open_incident_info_view(client, body, respond)
         return
     action, *args = args
     match action:
@@ -317,270 +320,6 @@ def handle_update_status_command(
             body["channel_name"],
             user_id,
         )
-
-
-def open_incident_info_view(client: WebClient, body, ack: Ack, respond: Respond):
-    """Open the incident information view. This view displays the incident details where certain fields can be updated."""
-    incident = db_operations.get_incident_by_channel_id(body["channel_id"])
-    if not incident:
-        respond(
-            "This is command is only available in incident channels. No incident records found for this channel."
-        )
-        return
-    else:
-        deserialize = TypeDeserializer()
-        incident_data = {k: deserialize.deserialize(v) for k, v in incident.items()}
-        view = incident_information_view(Incident(**incident_data))
-        client.views_open(trigger_id=body["trigger_id"], view=view)
-
-
-def open_update_field_view(client: WebClient, body, ack: Ack, respond: Respond):
-    """Open the view to update the incident field"""
-    ack()
-    action = body["actions"][0]["value"]
-    incident_data = json.loads(body["view"]["private_metadata"])
-    view = update_field_view(action, incident_data)
-    client.views_push(
-        view_id=body["view"]["id"], view=view, trigger_id=body["trigger_id"]
-    )
-
-
-def incident_information_view(incident: Incident):
-    """Create the view for the incident information modal.
-    It should receive a valid Incident object"""
-    created_at = "Unknown"
-    impact_start_timestamp = "Unknown"
-    impact_end_timestamp = "Unknown"
-    detection_timestamp = "Unknown"
-    if incident.created_at:
-        created_at = convert_timestamp(incident.created_at)
-    if incident.start_impact_time:
-        impact_start_timestamp = convert_timestamp(incident.start_impact_time)
-    if incident.end_impact_time:
-        impact_end_timestamp = convert_timestamp(incident.end_impact_time)
-    if incident.detection_time:
-        detection_timestamp = convert_timestamp(incident.detection_time)
-
-    incident_data = incident.model_dump()
-    private_metadata = json.dumps(incident_data)
-
-    return {
-        "type": "modal",
-        "callback_id": "incident_information_view",
-        "title": {"type": "plain_text", "text": "Incident Information", "emoji": True},
-        "close": {"type": "plain_text", "text": "OK", "emoji": True},
-        "private_metadata": private_metadata,
-        "blocks": [
-            {
-                "type": "header",
-                "text": {
-                    "type": "plain_text",
-                    "text": incident.name,
-                    "emoji": True,
-                },
-            },
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "*ID*: " + incident.id,
-                },
-            },
-            {"type": "divider"},
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "*Status*:\n" + incident.status,
-                },
-                "accessory": {
-                    "type": "button",
-                    "text": {"type": "plain_text", "text": "Update", "emoji": True},
-                    "value": "status",
-                    "action_id": "update_incident_field",
-                },
-            },
-            {"type": "divider"},
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "*Time Created*:\n" + created_at,
-                },
-            },
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "*Detection Time*:\n" + detection_timestamp,
-                },
-                "accessory": {
-                    "type": "button",
-                    "text": {"type": "plain_text", "text": "Update", "emoji": True},
-                    "value": "detection_time",
-                    "action_id": "update_incident_field",
-                },
-            },
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "*Start of Impact*:\n" + impact_start_timestamp,
-                },
-                "accessory": {
-                    "type": "button",
-                    "text": {"type": "plain_text", "text": "Update", "emoji": True},
-                    "value": "start_impact_time",
-                    "action_id": "update_incident_field",
-                },
-            },
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "*End of Impact*:\n" + impact_end_timestamp,
-                },
-                "accessory": {
-                    "type": "button",
-                    "text": {"type": "plain_text", "text": "Update", "emoji": True},
-                    "value": "end_impact_time",
-                    "action_id": "update_incident_field",
-                },
-            },
-        ],
-    }
-
-
-def update_field_view(action, incident_data):
-    logging.info("Loading Update Field View for action: %s", action)
-    date_actions = {
-        "detection_time",
-        "start_impact_time",
-        "end_impact_time",
-    }
-    if action in date_actions:
-        now = datetime.now()
-        if incident_data[action] != "Unknown":
-            now = datetime.fromtimestamp(float(incident_data[action]))
-        initial_date = now.strftime("%Y-%m-%d")
-        initial_time = now.strftime("%H:%M")
-        return {
-            "type": "modal",
-            "callback_id": "update_field_modal",
-            "title": {
-                "type": "plain_text",
-                "text": "Incident Information",
-                "emoji": True,
-            },
-            "submit": {"type": "plain_text", "text": "Submit", "emoji": True},
-            "close": {"type": "plain_text", "text": "OK", "emoji": True},
-            "private_metadata": json.dumps(
-                {"action": action, "incident_data": incident_data}
-            ),
-            "blocks": [
-                {
-                    "type": "header",
-                    "text": {
-                        "type": "plain_text",
-                        "text": action,
-                        "emoji": True,
-                    },
-                },
-                {
-                    "type": "input",
-                    "block_id": "date_input",
-                    "element": {
-                        "type": "datepicker",
-                        "initial_date": initial_date,
-                        "placeholder": {
-                            "type": "plain_text",
-                            "text": "Select a date",
-                        },
-                        "action_id": "date_picker",
-                    },
-                    "label": {
-                        "type": "plain_text",
-                        "text": "Select a date",
-                    },
-                },
-                {
-                    "type": "input",
-                    "block_id": "time_input",
-                    "element": {
-                        "type": "timepicker",
-                        "initial_time": initial_time,
-                        "placeholder": {
-                            "type": "plain_text",
-                            "text": "Select time",
-                            "emoji": True,
-                        },
-                        "action_id": "time_picker",
-                    },
-                    "label": {
-                        "type": "plain_text",
-                        "text": "Select time",
-                        "emoji": True,
-                    },
-                },
-            ],
-        }
-    else:
-        return {
-            "type": "modal",
-            "title": {
-                "type": "plain_text",
-                "text": "Incident Information",
-                "emoji": True,
-            },
-            "close": {"type": "plain_text", "text": "OK", "emoji": True},
-            "blocks": [
-                {
-                    "type": "header",
-                    "text": {
-                        "type": "plain_text",
-                        "text": action,
-                        "emoji": True,
-                    },
-                }
-            ],
-        }
-
-
-def handle_update_field_submission(client: WebClient, body, ack: Ack, view, logger):
-    ack()
-    user_id = body["user"]["id"]
-
-    private_metadata = json.loads(view["private_metadata"])
-    action = private_metadata["action"]
-    incident_data = private_metadata["incident_data"]
-    incident_id = incident_data["id"]
-    channel_id = incident_data["channel_id"]
-
-    date = view["state"]["values"]["date_input"]["date_picker"]["selected_date"]
-    time = view["state"]["values"]["time_input"]["time_picker"]["selected_time"]
-    timestamp = datetime.strptime(f"{date} {time}", "%Y-%m-%d %H:%M").timestamp()
-
-    field_map = {
-        "detection_time",
-        "start_impact_time",
-        "end_impact_time",
-    }
-
-    if action in field_map:
-        db_operations.update_incident_field(
-            logger, incident_id, action, str(timestamp), user_id, type="S"
-        )
-        client.chat_postMessage(
-            channel=channel_id,
-            text=f"<@{user_id}> has updated the field {action} to {date} {time}",
-        )
-
-        incident_data[action] = str(timestamp)
-        view = incident_information_view(Incident(**incident_data))
-        client.views_update(view_id=body["view"]["root_view_id"], view=view)
-
-    else:
-        logger.error("Unknown action: %s", action)
 
 
 def parse_incident_datetime_string(datetime_string: str) -> str:
