+++ app/modules/incident/information_display.py
@@ -1,0 +1,131 @@
+import json
+from boto3.dynamodb.types import TypeDeserializer
+from slack_bolt import Respond
+from slack_sdk import WebClient
+from models.incidents import Incident
+from modules.incident import db_operations, incident_helper
+
+
+def open_incident_info_view(client: WebClient, body, respond: Respond):
+    """Open the incident information view. This view displays the incident details where certain fields can be updated."""
+    incident = db_operations.get_incident_by_channel_id(body["channel_id"])
+    if not incident:
+        respond(
+            "This is command is only available in incident channels. No incident records found for this channel."
+        )
+        return
+    else:
+        deserialize = TypeDeserializer()
+        incident_data = {k: deserialize.deserialize(v) for k, v in incident.items()}
+        view = incident_information_view(Incident(**incident_data))
+        client.views_open(trigger_id=body["trigger_id"], view=view)
+
+
+def incident_information_view(incident: Incident):
+    """Create the view for the incident information modal.
+    It should receive a valid Incident object"""
+    created_at = "Unknown"
+    impact_start_timestamp = "Unknown"
+    impact_end_timestamp = "Unknown"
+    detection_timestamp = "Unknown"
+    if incident.created_at:
+        created_at = incident_helper.convert_timestamp(incident.created_at)
+    if incident.start_impact_time:
+        impact_start_timestamp = incident_helper.convert_timestamp(
+            incident.start_impact_time
+        )
+    if incident.end_impact_time:
+        impact_end_timestamp = incident_helper.convert_timestamp(
+            incident.end_impact_time
+        )
+    if incident.detection_time:
+        detection_timestamp = incident_helper.convert_timestamp(incident.detection_time)
+
+    incident_data = incident.model_dump()
+    private_metadata = json.dumps(incident_data)
+
+    return {
+        "type": "modal",
+        "callback_id": "incident_information_view",
+        "title": {"type": "plain_text", "text": "Incident Information", "emoji": True},
+        "close": {"type": "plain_text", "text": "OK", "emoji": True},
+        "private_metadata": private_metadata,
+        "blocks": [
+            {
+                "type": "header",
+                "text": {
+                    "type": "plain_text",
+                    "text": incident.name,
+                    "emoji": True,
+                },
+            },
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": "*ID*: " + incident.id,
+                },
+            },
+            {"type": "divider"},
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": "*Status*:\n" + incident.status,
+                },
+                "accessory": {
+                    "type": "button",
+                    "text": {"type": "plain_text", "text": "Update", "emoji": True},
+                    "value": "status",
+                    "action_id": "update_incident_field",
+                },
+            },
+            {"type": "divider"},
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": "*Time Created*:\n" + created_at,
+                },
+            },
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": "*Detection Time*:\n" + detection_timestamp,
+                },
+                "accessory": {
+                    "type": "button",
+                    "text": {"type": "plain_text", "text": "Update", "emoji": True},
+                    "value": "detection_time",
+                    "action_id": "update_incident_field",
+                },
+            },
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": "*Start of Impact*:\n" + impact_start_timestamp,
+                },
+                "accessory": {
+                    "type": "button",
+                    "text": {"type": "plain_text", "text": "Update", "emoji": True},
+                    "value": "start_impact_time",
+                    "action_id": "update_incident_field",
+                },
+            },
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": "*End of Impact*:\n" + impact_end_timestamp,
+                },
+                "accessory": {
+                    "type": "button",
+                    "text": {"type": "plain_text", "text": "Update", "emoji": True},
+                    "value": "end_impact_time",
+                    "action_id": "update_incident_field",
+                },
+            },
+        ],
+    }
