+++ app/tests/modules/incident/test_incident_helper.py
@@ -6,13 +6,12 @@ from unittest.mock import ANY, MagicMock, patch
 import pytest
 
 from modules import incident_helper
-from models.incidents import Incident
 
 SLACK_SECURITY_USER_GROUP_ID = os.getenv("SLACK_SECURITY_USER_GROUP_ID")
 
 
 @patch("modules.incident.incident_helper.incident_conversation")
-@patch("modules.incident.incident_helper.open_incident_info_view")
+@patch("modules.incident.incident_helper.information_display.open_incident_info_view")
 def test_handle_incident_command_with_empty_args(
     mock_open_incident_info_view, mock_incident_conversation
 ):
@@ -25,7 +24,7 @@ def test_handle_incident_command_with_empty_args(
     }
     mock_incident_conversation.is_incident_channel.return_value = (True, False)
     incident_helper.handle_incident_command([], client, body, respond, ack, logger)
-    mock_open_incident_info_view.assert_called_once_with(client, body, ack, respond)
+    mock_open_incident_info_view.assert_called_once_with(client, body, respond)
 
 
 @patch("modules.incident.incident_helper.google_drive.create_folder")
@@ -722,410 +721,6 @@ def test_handle_update_status_command_no_incidents_found(
         "No incident found for this channel. Will not update status in DB record."
     )
     mock_incident_status.update_status.assert_not_called()
-
-
-@patch("modules.incident.incident_helper.incident_information_view")
-@patch("modules.incident.incident_helper.db_operations")
-def test_open_incident_info_view(mock_db_operations, mock_incident_information_view):
-    mock_client = MagicMock()
-    mock_ack = MagicMock()
-    mock_respond = MagicMock()
-    body = {
-        "channel_id": "C12345",
-        "channel_name": "incident-2024-01-12-test",
-        "user_id": "U12345",
-        "trigger_id": "T12345",
-        "view": {"id": "V12345"},
-    }
-    mock_db_operations.get_incident_by_channel_id.return_value = {
-        "id": {"S": "incident-2024-01-12-test"},
-        "channel_id": {"S": "C1234567890"},
-        "channel_name": {"S": "incident-2024-01-12-test"},
-        "name": {"S": "Test Incident"},
-        "user_id": {"S": "U12345"},
-        "teams": {"L": [{"S": "team1"}, {"S": "team2"}]},
-        "report_url": {"S": "http://example.com/report"},
-        "status": {"S": "Open"},
-        "meet_url": {"S": "http://example.com/meet"},
-        "created_at": {"S": "1234567890"},
-        "incident_commander": {"S": "Commander"},
-        "operations_lead": {"S": "Lead"},
-        "severity": {"S": "High"},
-        "start_impact_time": {"S": "1234567890"},
-        "end_impact_time": {"S": "1234567890"},
-        "detection_time": {"S": "1234567890"},
-        "retrospective_url": {"S": "http://example.com/retrospective"},
-        "environment": {"S": "prod"},
-        "logs": {"L": []},
-        "incident_updates": {"L": []},
-    }
-
-    incident_data = {
-        "id": "incident-2024-01-12-test",
-        "channel_id": "C1234567890",
-        "channel_name": "incident-2024-01-12-test",
-        "name": "Test Incident",
-        "user_id": "U12345",
-        "teams": ["team1", "team2"],
-        "report_url": "http://example.com/report",
-        "status": "Open",
-        "meet_url": "http://example.com/meet",
-        "created_at": "1234567890",
-        "incident_commander": "Commander",
-        "operations_lead": "Lead",
-        "severity": "High",
-        "start_impact_time": "1234567890",
-        "end_impact_time": "1234567890",
-        "detection_time": "1234567890",
-        "retrospective_url": "http://example.com/retrospective",
-        "environment": "prod",
-        "logs": [],
-        "incident_updates": [],
-    }
-
-    mock_incident_information_view.return_value = {"view": [{"block": "block_id"}]}
-    incident_helper.open_incident_info_view(mock_client, body, mock_ack, mock_respond)
-    mock_client.views_open.assert_called_once_with(
-        trigger_id="T12345",
-        view={"view": [{"block": "block_id"}]},
-    )
-    mock_incident_information_view.assert_called_once_with(Incident(**incident_data))
-
-
-@patch("modules.incident.incident_helper.db_operations")
-def test_open_incident_view_no_incident_found(mock_db_operations):
-    mock_client = MagicMock()
-    mock_ack = MagicMock()
-    mock_respond = MagicMock()
-    body = {
-        "channel_id": "C12345",
-        "channel_name": "incident-2024-01-12-test",
-        "user_id": "U12345",
-        "trigger_id": "T12345",
-        "view": {"id": "V12345"},
-    }
-    mock_db_operations.get_incident_by_channel_id.return_value = []
-    incident_helper.open_incident_info_view(mock_client, body, mock_ack, mock_respond)
-    mock_respond.assert_called_once_with(
-        "This is command is only available in incident channels. No incident records found for this channel."
-    )
-
-
-@patch("modules.incident.incident_helper.logging")
-@patch("modules.incident.incident_helper.update_field_view")
-def test_open_update_field_view(mock_update_field_view, mock_logging):
-    mock_client = MagicMock()
-    mock_ack = MagicMock()
-    mock_respond = MagicMock()
-    private_metadata = json.dumps({"status": "data"})
-    body = {
-        "channel_id": "C12345",
-        "channel_name": "incident-2024-01-12-test",
-        "user_id": "U12345",
-        "trigger_id": "T12345",
-        "view": {"id": "V12345", "private_metadata": private_metadata},
-        "actions": [
-            {"action_id": "update_incident_field", "value": "action_to_perform"}
-        ],
-    }
-    mock_update_field_view.return_value = {"view": [{"block": "block_id"}]}
-    incident_helper.open_update_field_view(mock_client, body, mock_ack, mock_respond)
-    mock_update_field_view.assert_called_once_with(
-        "action_to_perform", {"status": "data"}
-    )
-    mock_client.views_push.assert_called_once_with(
-        trigger_id="T12345",
-        view_id="V12345",
-        view={"view": [{"block": "block_id"}]},
-    )
-
-
-@patch("modules.incident.incident_helper.convert_timestamp")
-@patch("modules.incident.incident_helper.logging")
-def test_incident_information_view(mock_logging, mock_convert_timestamp):
-    incident_data = generate_incident_data(
-        start_impact_time="1234567890",
-        end_impact_time="1234567890",
-        detection_time="1234567890",
-    )
-    id = incident_data["id"]
-    mock_convert_timestamp.side_effect = [
-        "2009-02-13 23:31:30",
-        "Unknown",
-        "Unknown",
-        "Unknown",
-    ]
-    incident = Incident(**incident_data)
-    private_metadata = json.dumps(incident.model_dump())
-    view = incident_helper.incident_information_view(incident)
-    assert view == {
-        "type": "modal",
-        "callback_id": "incident_information_view",
-        "title": {
-            "type": "plain_text",
-            "text": "Incident Information",
-            "emoji": True,
-        },
-        "close": {"type": "plain_text", "text": "OK", "emoji": True},
-        "private_metadata": private_metadata,
-        "blocks": [
-            {
-                "type": "header",
-                "text": {
-                    "type": "plain_text",
-                    "text": "name",
-                    "emoji": True,
-                },
-            },
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "*ID*: " + id,
-                },
-            },
-            {"type": "divider"},
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "*Status*:\nstatus",
-                },
-                "accessory": {
-                    "type": "button",
-                    "text": {"type": "plain_text", "text": "Update", "emoji": True},
-                    "value": "status",
-                    "action_id": "update_incident_field",
-                },
-            },
-            {"type": "divider"},
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "*Time Created*:\n2009-02-13 23:31:30",
-                },
-            },
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "*Detection Time*:\nUnknown",
-                },
-                "accessory": {
-                    "type": "button",
-                    "text": {"type": "plain_text", "text": "Update", "emoji": True},
-                    "value": "detection_time",
-                    "action_id": "update_incident_field",
-                },
-            },
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "*Start of Impact*:\nUnknown",
-                },
-                "accessory": {
-                    "type": "button",
-                    "text": {"type": "plain_text", "text": "Update", "emoji": True},
-                    "value": "start_impact_time",
-                    "action_id": "update_incident_field",
-                },
-            },
-            {
-                "type": "section",
-                "text": {
-                    "type": "mrkdwn",
-                    "text": "*End of Impact*:\nUnknown",
-                },
-                "accessory": {
-                    "type": "button",
-                    "text": {"type": "plain_text", "text": "Update", "emoji": True},
-                    "value": "end_impact_time",
-                    "action_id": "update_incident_field",
-                },
-            },
-        ],
-    }
-
-
-@patch("modules.incident.incident_helper.logging")
-def test_update_field_view(mock_logging):
-    view = incident_helper.update_field_view("status", {"status": "data"})
-    mock_logging.info.assert_called_once_with(
-        "Loading Update Field View for action: %s", "status"
-    )
-    assert view == {
-        "type": "modal",
-        "title": {"type": "plain_text", "text": "Incident Information", "emoji": True},
-        "close": {"type": "plain_text", "text": "OK", "emoji": True},
-        "blocks": [
-            {
-                "type": "header",
-                "text": {
-                    "type": "plain_text",
-                    "text": "status",
-                    "emoji": True,
-                },
-            }
-        ],
-    }
-
-
-@patch("modules.incident.incident_helper.logging")
-def test_update_field_view_date_field(mock_logging):
-    incident_data = {"status": "data", "detection_time": "1234567890"}
-    view = incident_helper.update_field_view("detection_time", incident_data)
-    mock_logging.info.assert_called_once_with(
-        "Loading Update Field View for action: %s", "detection_time"
-    )
-    assert view == {
-        "type": "modal",
-        "callback_id": "update_field_modal",
-        "title": {
-            "type": "plain_text",
-            "text": "Incident Information",
-            "emoji": True,
-        },
-        "submit": {"type": "plain_text", "text": "Submit", "emoji": True},
-        "close": {"type": "plain_text", "text": "OK", "emoji": True},
-        "private_metadata": json.dumps(
-            {"action": "detection_time", "incident_data": incident_data}
-        ),
-        "blocks": [
-            {
-                "type": "header",
-                "text": {
-                    "type": "plain_text",
-                    "text": "detection_time",
-                    "emoji": True,
-                },
-            },
-            {
-                "type": "input",
-                "block_id": "date_input",
-                "element": {
-                    "type": "datepicker",
-                    "initial_date": "2009-02-13",
-                    "placeholder": {
-                        "type": "plain_text",
-                        "text": "Select a date",
-                    },
-                    "action_id": "date_picker",
-                },
-                "label": {
-                    "type": "plain_text",
-                    "text": "Select a date",
-                },
-            },
-            {
-                "type": "input",
-                "block_id": "time_input",
-                "element": {
-                    "type": "timepicker",
-                    "initial_time": "23:31",
-                    "placeholder": {
-                        "type": "plain_text",
-                        "text": "Select time",
-                        "emoji": True,
-                    },
-                    "action_id": "time_picker",
-                },
-                "label": {
-                    "type": "plain_text",
-                    "text": "Select time",
-                    "emoji": True,
-                },
-            },
-        ],
-    }
-
-
-@patch("modules.incident.incident_helper.incident_information_view")
-@patch("modules.incident.incident_helper.db_operations")
-def test_handle_update_field_submission(
-    mock_db_operations, mock_incident_information_view
-):
-    mock_client = MagicMock()
-    mock_ack = MagicMock()
-    mock_logger = MagicMock()
-    # update value as a string based on date and time input 2024-01-12 12:00
-    updated_value = "1705060800.0"
-    incident_data = generate_incident_data()
-    view = {
-        "state": {
-            "values": {
-                "date_input": {"date_picker": {"selected_date": "2024-01-12"}},
-                "time_input": {"time_picker": {"selected_time": "12:00"}},
-            }
-        },
-        "private_metadata": json.dumps(
-            {
-                "action": "detection_time",
-                "incident_data": incident_data,
-            }
-        ),
-    }
-    body = {
-        "user": {"id": incident_data["user_id"]},
-        "view": {"root_view_id": "root_view_id"},
-    }
-    mock_incident_information_view.return_value = {"view": [{"block": "block_id"}]}
-
-    incident_helper.handle_update_field_submission(
-        mock_client, body, mock_ack, view, mock_logger
-    )
-    mock_db_operations.update_incident_field.assert_called_once_with(
-        mock_logger,
-        incident_data["id"],
-        "detection_time",
-        updated_value,
-        incident_data["user_id"],
-        type="S",
-    )
-    mock_client.chat_postMessage.assert_called_once_with(
-        channel=incident_data["channel_id"],
-        text="<@user_id> has updated the field detection_time to 2024-01-12 12:00",
-    )
-
-
-@patch("modules.incident.incident_helper.incident_information_view")
-@patch("modules.incident.incident_helper.db_operations")
-def test_handle_update_field_submission_not_supported(
-    mock_db_operations, mock_incident_information_view
-):
-    mock_client = MagicMock()
-    mock_ack = MagicMock()
-    mock_logger = MagicMock()
-    incident_data = generate_incident_data()
-    view = {
-        "state": {
-            "values": {
-                "date_input": {"date_picker": {"selected_date": "2024-01-12"}},
-                "time_input": {"time_picker": {"selected_time": "12:00"}},
-            }
-        },
-        "private_metadata": json.dumps(
-            {
-                "action": "unsupported_field",
-                "incident_data": incident_data,
-            }
-        ),
-    }
-    body = {
-        "user": {"id": incident_data["user_id"]},
-        "view": {"root_view_id": "root_view_id"},
-    }
-    mock_incident_information_view.return_value = {"view": [{"block": "block_id"}]}
-
-    incident_helper.handle_update_field_submission(
-        mock_client, body, mock_ack, view, mock_logger
-    )
-    mock_db_operations.update_incident_field.assert_not_called()
-    mock_client.chat_postMessage.assert_not_called()
-    mock_incident_information_view.assert_not_called()
-    mock_client.views_update.assert_not_called()
-    mock_logger.error.assert_called_once_with("Unknown action: %s", "unsupported_field")
 
 
 def test_parse_incident_datetime_string():
